+++ R/fromnamespace.R
@@ -30,4 +30,3 @@ with_seed_null <- function(seed, code) {
     }
 }
 
-##' @import cli